Teaching_Agent/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
chromadb>=0.4.0
pdfplumber>=0.10.0
questionary>=2.0.0
tqdm>=4.0.0